build: make sysroot-linux contain libstdc++ and gcc stubs
This is a little irritating, but we need two deps that _should_
be covered by a hermetic all LLVM toolchain.

The first is `libgcc_s.so`, this is part of the GCC compiler 
runtime. In this case this is down to rustc assuming a gnu abi
mandates this library for the requirement of unwinding. This is 
a little difficult to avoid without bootstraping rustc which we
avvoid for now.

The second is `libstdc++.so`, this is the C++ standard library
expressed as part of GNU. We are right now avoiding this as its
easier from an ABI and language version standpoint to use libc++
however we want to pick up torch and ray which presently in the
prepackaged versions are compiled with dynmic linking to libstdc++